allZeroLag

+++ internal/gameServer/udp.go
@@ -287,5 +287,5 @@ func (g *GameServer) createUDPServer() error {
 }
 func (g *GameServer) updateBufferSize(oldBufferInt int32, i int) {
 	// Set the specific buffer index i to the oldBufferInt value
-	g.GameData.BufferSize[i] = oldBufferInt
+	g.GameData.BufferSize[i] = uint32(oldBufferInt)
 }
